refactor: break out QueryPluginsService from PluginInformationService
